+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added Estonian Language Support
 - Added Greek Language Support
 - Added Croatian language support
+- Added Icelandic language support
 
 ## [4.18.0] - 2025-09-xx
 ### Added